+++ packages/vip-go/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.1](https://github.com/alleyinteractive/irving/packages/vip-go/compare/v2.11.0...v2.11.1) (2020-06-03)
+
+**Note:** Version bump only for package @irvingjs/vip-go
+
+
+
+
+
+# [2.11.0](https://github.com/alleyinteractive/irving/packages/vip-go/compare/v2.11.0-rc.0...v2.11.0) (2020-06-02)
+
+**Note:** Version bump only for package @irvingjs/vip-go
+
+
+
+
+
+# [2.11.0-rc.0](https://github.com/alleyinteractive/irving/packages/vip-go/compare/v2.11.0-beta.22...v2.11.0-rc.0) (2020-06-02)
+
+**Note:** Version bump only for package @irvingjs/vip-go
+
+
+
+
+
+# [2.11.0-beta.21](https://github.com/alleyinteractive/irving/packages/vip-go/compare/v2.11.0-beta.20...v2.11.0-beta.21) (2020-06-02)
+
+
+### Features
+
+* **cache:** separate client and service creation for cache ([5a40671](https://github.com/alleyinteractive/irving/packages/vip-go/commit/5a40671ede6e51eb9b84132845d224d5601621fa))
+
+
+
+
+
+# [2.11.0-beta.20](https://github.com/alleyinteractive/irving/packages/vip-go/compare/v2.11.0-beta.19...v2.11.0-beta.20) (2020-06-02)
+
+**Note:** Version bump only for package @irvingjs/vip-go
+
+
+
+
+
+# [2.11.0-beta.19](https://github.com/alleyinteractive/irving/packages/vip-go/compare/v2.11.0-beta.18...v2.11.0-beta.19) (2020-06-02)
+
+**Note:** Version bump only for package @irvingjs/vip-go
+
+
+
+
+
+# [2.11.0-beta.18](https://github.com/alleyinteractive/irving/packages/vip-go/compare/v2.11.0-beta.17...v2.11.0-beta.18) (2020-06-02)
+
+**Note:** Version bump only for package @irvingjs/vip-go
+
+
+
+
+
 # [2.11.0-beta.17](https://github.com/alleyinteractive/irving/packages/vip-go/compare/v2.11.0-beta.16...v2.11.0-beta.17) (2020-05-28)
 
 **Note:** Version bump only for package @irvingjs/vip-go